pyproject.toml
[tool.poetry]
name = "nonebot-plugin-nicklock"
version = "0.1.0"
description = "A nonebot plugin to automatically reset your group nickname."
authors = ["KoishiMoe <68314080+KoishiMoe@users.noreply.github.com>"]
license = "AGPL-3.0-or-later"
readme = "README.md"
repository = "https://github.com/KoishiMoe/nonebot-plugin-nicklock"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = ">=2.0.0b3"
nonebot-adapter-onebot = ">=2.0.0.b1"
nonebot-plugin-apscheduler = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
